model: add sqlite3 support
To build with sqlite3, we need to rearrange the source a bit to make
'swift build' happy.  This needs a manual argument for sqlite3, so
create a build.sh script to invoke the build correctly.

The sqlite3 code creates a simple database on run, and populates it
with the problems read from the json file.