Use a database with a fixed path
When the app starts, create or open a database with a fixed path.
`~/Documents/Playful.sqlite`.  This will ultimately need to be
reworked to use a model that fits better with MacOS conventions, but
this will let us get a main part of the GUI working without needing to
figure out how to do it properly.